Move build_witness into a fn
